requirements.txt
absl-py==0.15.0
asgiref==3.4.1
astunparse==1.6.3
beautifulsoup4==4.11.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.0.12
clang==5.0
dataclasses==0.8
dj-database-url==0.5.0
Django==3.2.13
django-bootstrap4==21.1
djangorestframework==3.13.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.1.0
idna==3.3
importlib-metadata==2.1.3
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.6
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.15.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.9
Werkzeug==2.0.3
whitenoise==5.3.0
wrapt==1.12.1
zipp==3.6.0
